requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pydantic==2.4.0
httpx==0.26.3
python-dotenv==1.0.0
prometheus-client==0.18.0
slowapi==0.1.4
redis==4.5.5
kafka-python==2.0.2
pydantic-settings==2.0.0
PyJWT==2.8.0
